добавил роли в виде админа/суперадмина/обычного пользователя, логика пользователей, измененные тесты(падают), измененная модель роли

+++ api/handlers.py
@@ -7,6 +7,7 @@ from fastapi import HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from api.actions.user import check_user_permissions
 from api.actions.auth import get_current_user_from_token
 from api.actions.user import _create_new_user
 from api.actions.user import _delete_user
@@ -34,12 +35,23 @@ async def create_user(body: UserCreate, db: AsyncSession = Depends(get_db)) -> S
         raise HTTPException(status_code=503, detail=f"Database error: {err}")
 
 
+
 @user_router.delete("/", response_model=DeleteUserResponse)
 async def delete_user(
         user_id: UUID,
         db: AsyncSession = Depends(get_db),
         current_user: User = Depends(get_current_user_from_token),
 ) -> DeleteUserResponse:
+    user_for_deletion = await _get_user_by_id(user_id, db)
+    if user_for_deletion is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with {user_id} not found."
+        )
+    if not check_user_permissions(
+        target_user=user_for_deletion,
+        current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
     deleted_user_id = await _delete_user(user_id, db)
     if deleted_user_id is None:
         raise HTTPException(
@@ -75,11 +87,15 @@ async def update_user_by_id(
             status_code=422,
             detail="At least one parameter for user update info should be provided",
         )
-    user = await _get_user_by_id(user_id, db)
-    if user is None:
+    user_for_update = await _get_user_by_id(user_id, db)
+    if user_for_update is None:
         raise HTTPException(
             status_code=404, detail=f"User with id {user_id} not found."
         )
+    if check_user_permissions(
+        target_user=user_for_update, current_user=current_user
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
     try:
         updated_user_id = await _update_user(
             updated_user_params=updated_user_params, session=db, user_id=user_id
